Refactor some functions into util/util.js.
